# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bella API",
    "description": "Backend for Bella \u2014 a personal AI companion.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/session": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange a Firebase ID token for a Bella session",
        "operationId": "create_session_v1_auth_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Revoke this session",
        "operationId": "logout_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BellaAccessToken": []
          }
        ]
      }
    },
    "/v1/auth/logout-all": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Revoke every session for this user",
        "operationId": "logout_all_v1_auth_logout_all_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BellaAccessToken": []
          }
        ]
      }
    },
    "/v1/auth/sessions": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List active sessions",
        "operationId": "list_sessions_v1_auth_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BellaAccessToken": []
          }
        ]
      }
    },
    "/v1/auth/sessions/{session_id}": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Revoke one session by id",
        "operationId": "revoke_session_v1_auth_sessions__session_id__delete",
        "security": [
          {
            "BellaAccessToken": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthenticatedUser": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_id": {
            "type": "string",
            "title": "Display Id"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "onboarding_complete": {
            "type": "boolean",
            "title": "Onboarding Complete"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_id",
          "preferred_name",
          "role",
          "locale",
          "timezone",
          "onboarding_complete"
        ],
        "title": "AuthenticatedUser"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RevocationResponse": {
        "properties": {
          "revoked": {
            "type": "integer",
            "title": "Revoked"
          }
        },
        "type": "object",
        "required": [
          "revoked"
        ],
        "title": "RevocationResponse"
      },
...
      "SessionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SessionListResponse"
      },
      "SessionRequest": {
        "properties": {
          "id_token": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Id Token"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          }
        },
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "SessionRequest"
      },
      "SessionResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          },
          "flags": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Flags"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "SessionResponse"
      },
      "SessionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "current": {
            "type": "boolean",
            "title": "Current"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_name",
          "created_at",
          "last_used_at",
          "expires_at",
          "current"
        ],
        "title": "SessionSummary",
        "description": "One active session, for the \"where am I signed in?\" screen."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BellaAccessToken": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}